FinalVersionStart
a rough version of what the final application will look like  currently
working on it more.  Some stuff is just filler so that css can be
experimented on it? idk how css works in qt so I just left some of the
code from what I copy and pasted so that there's widgets there but now
that I know we're dealing with local files I can just start making the
actual widgets now.  Hope to be done by the end of the weekend.