Write Up
This is what I have so far. I added in a table for concepts we were
supposed to demonstrate, if there was some stuff I missed (which I
probably did), feel free to let me know, or add it in yourself. This
isn't exactly finished, but it should work if we had to turn it in right
now.